projects/trading/requirements.txt
duckdb>=0.9.0
yfinance>=0.2.30
pandas>=2.1.0
jupyterlab>=4.0.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
